requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.35.0
pydantic-settings==2.10.1
email-validator==2.2.0
PyJWT[crypto]==2.9.0
supabase==2.10.0
